Crate storage_service_server

Modules

network

Structs

Handler

The Handler is the “pure” inbound request handler. It contains all the necessary context and state needed to construct a response to an inbound request. We usually clone/create a new handler for every request.

StorageReader

The underlying implementation of the StorageReaderInterface, used by the storage server.

StorageServiceServer

The server-side actor for the storage service. Handles inbound storage service requests from clients.

Enums

Error

Constants

STORAGE_SERVER_VERSION

Storage server constants.

Traits

StorageReaderInterface

The interface into local storage (e.g., the Diem DB) used by the storage server to handle client requests.