Trait storage_service_server::StorageReaderInterface

pub trait StorageReaderInterface {
    fn get_data_summary(&self) -> Result<DataSummary, Error>;

    fn get_transactions_with_proof(
        &self, 
        proof_version: u64, 
        start_version: u64, 
        expected_num_transactions: u64, 
        include_events: bool
    ) -> Result<TransactionListWithProof, Error>;

    fn get_epoch_ending_ledger_infos(
        &self, 
        start_epoch: u64, 
        expected_end_epoch: u64
    ) -> Result<EpochChangeProof, Error>;

    fn get_transaction_outputs_with_proof(
        &self, 
        proof_version: u64, 
        start_version: u64, 
        expected_num_transaction_outputs: u64
    ) -> Result<TransactionOutputListWithProof, Error>;

    fn get_account_states_chunk_with_proof(
        &self, 
        version: u64, 
        start_account_key: HashValue, 
        expected_num_account_states: u64
    ) -> Result<AccountStatesChunkWithProof, Error>;
}

The interface into local storage (e.g., the Diem DB) used by the storage server to handle client requests.

Required methods

fn get_data_summary(&self) -> Result<DataSummary, Error>

Returns a data summary of the underlying storage state.

fn get_transactions_with_proof(
&self,
proof_version: u64,
start_version: u64,
expected_num_transactions: u64,
include_events: bool
) -> Result<TransactionListWithProof, Error>

Returns a list of transactions with a proof relative to the proof_version. The transaction list is expected to contain at most expected_num_transactions and start at start_version. If include_events is true, events are also returned.

fn get_epoch_ending_ledger_infos(
&self,
start_epoch: u64,
expected_end_epoch: u64
) -> Result<EpochChangeProof, Error>

Returns a list of epoch ending ledger infos, starting at start_epoch and ending at most at the expected_end_epoch.

fn get_transaction_outputs_with_proof(
&self,
proof_version: u64,
start_version: u64,
expected_num_transaction_outputs: u64
) -> Result<TransactionOutputListWithProof, Error>

Returns a list of transaction outputs with a proof relative to the proof_version. The transaction output list is expected to contain at most expected_num_transaction_outputs and start at start_version.

fn get_account_states_chunk_with_proof(
&self,
version: u64,
start_account_key: HashValue,
expected_num_account_states: u64
) -> Result<AccountStatesChunkWithProof, Error>

Returns a chunk holding a list of account states starting at the specified account key with at most expected_num_account_states.